requirements.txt
python-telegram-bot==21.5
sympy==1.13.1
python-dotenv==1.0.1
